+++ src/interfaces/user.ts
@@ -1,16 +1,17 @@
 export interface IUser {
-  UNICode?       : string
-  condition      : string
-  documentNumber?: string
-  documentType?  : string
-  faculty?       : string
-  id?            : string
-  lastName?      : string
-  mail?          : string
-  names?         : string
-  optionalMail?  : string
-  postulating?   : boolean
-  registered?    : boolean
-  secondLastName?: string
-  specialty?     : string
+  UNICode?        : string
+  committeeMember?: boolean
+  condition?      : string
+  documentNumber? : string
+  documentType?   : string
+  faculty?        : string
+  id?             : string
+  lastName?       : string
+  mail?           : string
+  names?          : string
+  optionalMail?   : string
+  postulating?    : boolean
+  registered?     : boolean
+  secondLastName? : string
+  specialty?      : string
 }